airflow/requirements.txt
apache-airflow==2.9.0
pandas
psycopg2-binary
clickhouse-driver
airflow-clickhouse-plugin